refactor: Move types to separate module
