+++ src/course/course-access.api.mock.ts
@@ -1,15 +1,4 @@
-export type CourseRole = 'student' | 'teacher';
-
-export type CourseAccessResult =
-  | {
-      status: 'allowed';
-      courseSlug: string;
-      username: string;
-      role: CourseRole;
-      studentUsernames: string[];
-    }
-  | { status: 'course-not-found'; courseSlug: string; username: string }
-  | { status: 'not-course-participant'; courseSlug: string; username: string };
+import type { CourseAccessResult } from './course-access.types';
 
 type MockCourseAccess = {
   teachers: string[];
